+++ Changelog.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   #199 by @MaxDesiatov and @mattt.
 - Added `--minimum-access-level` option to `generate` and `coverage` commands.
   #219 by @Lukas-Stuehrk.
+- Added support for documenting default implementations.
+  #221 by @Lukas-Stuehrk.
 
 ### Fixed